Separate side effects from package entry point
