+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@unicef-polymer/etools-data-table",
   "description": "Data table",
-  "version": "4.0.1",
+  "version": "4.1.0-rc.1",
   "homepage": "https://github.com/unicef-polymer/etools-data-table",
   "repository": {
     "type": "git",
